fix(jdbc): refuse lossy float narrowing and redact at-signs in passwords
getFloat cast FalkorDB's 64-bit double straight to a float, so a value
above Float.MAX_VALUE came back as a fabricated infinity and a small
non-zero value as a fabricated zero. Every other narrowing getter refuses
a conversion that changes the value, so these now do too, through a
checked GraphValues.asFloat that the float[] unpacker also uses. A
vector's components are already 32-bit, so they stay exact.

The error redactor stopped the password at the first '@', while the URL
parser splits userinfo on the last one. A password containing '@' was
therefore only partly masked, leaking its tail into the SQLException
message. The password group now runs greedily to the same delimiter the
parser uses.

+++ src/main/java/com/falkordb/jdbc/FalkorDBResultSet.java
@@ -372,7 +372,7 @@ public final class FalkorDBResultSet extends FalkorDBWrapper implements ResultSe
 
     @Override
     public float getFloat(int columnIndex) throws SQLException {
-        return (float) GraphValues.asDouble(raw(columnIndex));
+        return GraphValues.asFloat(raw(columnIndex));
     }
 
     @Override
